wip: add collect() to multisearch components
