+++ gradle.properties
@@ -7,7 +7,7 @@ org.gradle.parallel=true
 
 # Lib
 GROUP=com.algolia
-VERSION_NAME=2.11.1
+VERSION_NAME=2.12.0-beta01
 
 # POM
 POM_DESCRIPTION=InstantSearch Android is a library providing widgets and helpers to help you build the best instant-search experience on Android with Algolia. It is built on top of Algolia's Kotlin API Client to provide you a high-level solution to quickly build various search interfaces.
